move PhoneAppPatches.cs into HomeScreen.Start.cs
